// szamlazzponthu-test-ui-app/src/app/core/services/api/models/Usr.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Usr API
 * CRUD műveletek a Usr erőforráson.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Usr
 */
export interface Usr {
    /**
     * 
     * @type {number}
     * @memberof Usr
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof Usr
     */
    firstname: string;
    /**
     * 
     * @type {string}
     * @memberof Usr
     */
    lastname: string;
    /**
     * 
     * @type {string}
     * @memberof Usr
     */
    address: string;
    /**
     * 
     * @type {string}
     * @memberof Usr
     */
    telephone: string;
    /**
     * 
     * @type {boolean}
     * @memberof Usr
     */
    active: boolean;
    /**
     * 
     * @type {string}
     * @memberof Usr
     */
    job: string;
}

/**
 * Check if a given object implements the Usr interface.
 */
export function instanceOfUsr(value: object): value is Usr {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('firstname' in value) || value['firstname'] === undefined) return false;
    if (!('lastname' in value) || value['lastname'] === undefined) return false;
    if (!('address' in value) || value['address'] === undefined) return false;
    if (!('telephone' in value) || value['telephone'] === undefined) return false;
    if (!('active' in value) || value['active'] === undefined) return false;
    if (!('job' in value) || value['job'] === undefined) return false;
    return true;
}

export function UsrFromJSON(json: any): Usr {
    return UsrFromJSONTyped(json, false);
}

export function UsrFromJSONTyped(json: any, ignoreDiscriminator: boolean): Usr {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'firstname': json['firstname'],
        'lastname': json['lastname'],
        'address': json['address'],
        'telephone': json['telephone'],
        'active': json['active'],
        'job': json['job'],
    };
}

export function UsrToJSON(json: any): Usr {
    return UsrToJSONTyped(json, false);
}

export function UsrToJSONTyped(value?: Usr | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'firstname': value['firstname'],
        'lastname': value['lastname'],
        'address': value['address'],
        'telephone': value['telephone'],
        'active': value['active'],
        'job': value['job'],
    };
}