// szamlazzponthu-test-ui-app/src/app/core/services/api/models/JobType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Usr API
 * CRUD műveletek a Usr erőforráson.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface JobType
 */
export interface JobType {
    /**
     * 
     * @type {string}
     * @memberof JobType
     */
    code: string;
    /**
     * 
     * @type {string}
     * @memberof JobType
     */
    labelHu: string;
    /**
     * 
     * @type {string}
     * @memberof JobType
     */
    labelEn: string;
}

/**
 * Check if a given object implements the JobType interface.
 */
export function instanceOfJobType(value: object): value is JobType {
    if (!('code' in value) || value['code'] === undefined) return false;
    if (!('labelHu' in value) || value['labelHu'] === undefined) return false;
    if (!('labelEn' in value) || value['labelEn'] === undefined) return false;
    return true;
}

export function JobTypeFromJSON(json: any): JobType {
    return JobTypeFromJSONTyped(json, false);
}

export function JobTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): JobType {
    if (json == null) {
        return json;
    }
    return {
        
        'code': json['code'],
        'labelHu': json['label_hu'],
        'labelEn': json['label_en'],
    };
}

export function JobTypeToJSON(json: any): JobType {
    return JobTypeToJSONTyped(json, false);
}

export function JobTypeToJSONTyped(value?: JobType | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'code': value['code'],
        'label_hu': value['labelHu'],
        'label_en': value['labelEn'],
    };
}